I moved the file to the folder textOptions
