+++ src/index.js
@@ -5,7 +5,7 @@ import { textOptions__process_layout__ } from "./layout";
 import { error__main_filter__ } from "./error";
 import { error__props_filter__ } from "./error/props-filter.error";
 
-import { process__SelectClass__ } from "./process/specialClass.process";
+import { process__SelectClass__ } from "./process/textOptions/specialClass.process";
 import { process__select_specialWrapp__ } from "./process/specialWrapp.process";
 
 import { whatItIs } from "./common/whatIsIt";
@@ -14,7 +14,7 @@ import { process__structure__ } from "./process/structure.process";
 import { memo__process } from "./process/memo.process";
 import { __Text__process__ } from "./process/text.class";
 import { __Structure__process__ } from "./process/structure.class";
-import { __TextOptions__process__ } from "./process/textOptions.class";
+import { __TextOptions__process__ } from "./process/textOptions/textOptions.class";
 
 function WL(
   props = {
