+++ src/process/textOptions/base.class.js
@@ -1,0 +1,3 @@
+export class Base {
+  constructor() {}
+}
