+++ src/process/textOptions/specialClass.process.js
