+++ src/process/textOptions/specialWrapp.process.js
